Added shadow copy to LCD. Now works with no LCD connected
The OLEDFourBit class now always keeps a copy up to date of the LCD
content. The content does not have to be read from the display anymore.

I also set a maximum wait time on the waitBusy function, so that it is
non blocking.

These two changes make it possible to run BrewPi without a display
connected. The display in the web interface still works.